fix(access-governance): address code review findings
Six fixes from a review of the whole branch, plus doc corrections.

Upload failures returned 500 instead of 400: multer.MulterError carries no
status property and this repo's fileFilter set none either, so both oversize
files and unsupported types fell through to the generic server-error handler.
Adds a router-level handler mirroring routes/evidence.js.

GET /entitlements is now paginated (page/limit, default 50, max 100). The
aggregates stay org-wide rather than per-page since they describe the whole
organization's posture; the response carries a pagination object and the UI
gains a pager plus a "showing N of M" line.

createCampaign batched into a single UNNEST insert instead of one round trip
per user, and simulateAccess now dedupes roleIds before comparing the matched
row count against the id count -- `= ANY(array)` is set membership, so a
repeated id previously produced a spurious 400.

The Jest decrypt mock was an identity function, so the suite passed whether
or not the service decrypted users.email -- the same blind spot that let
ciphertext reach the UI. It now returns distinguishable plaintext, and three
getEntitlementReport tests pin decryption and pagination. Each fix was
verified by reverting it and confirming the suite fails.

Extracts the Import & AI tab into components/accessGovernance/ (1179 -> 844
lines); components/rmf/ and components/ai/ already set that precedent here.

Adds backend/uploads/ to .gitignore: completeCampaign writes access review
summaries there containing subject and reviewer emails, and unlike the
sibling repo this one had no ignore rule for it.

Docs: adds docs/ACCESS_GOVERNANCE.md (none existed here), documents the
pagination behavior, and states the real scope of the self-review
restriction -- access_governance.manage ships only to admin, admins hold the
'*' wildcard, and requireSod treats it as an authorized override, so the
block does not fire in a default configuration. Migration 134's header now
records that RLS is inert for superuser connections.
