+++ docs/doc-review-log/PR-237.md
@@ -1,0 +1,98 @@
+# Doc Review — PR #237 (Access Governance module port)
+
+Reviewed at head commit `3d9794e`, covering the full branch: feature/port
+commit `5f44210`, dependency advisory fix `590ecea`, the decrypt + RLS
+security fix `ae99d29`, the compliance-labeler fix `3d9794e`, and this pass's
+code-review remediation.
+
+This repo has no automated docs-pipeline bot (unlike the sibling
+ControlWeaver-Pro repo), so per `.claude/rules/doc-review.md` the cue here was
+a meaningfully-sized diff touching behavior that top-level docs describe.
+
+## What the PR changes
+
+- Backend module `/api/v1/access-governance`: entitlement report, SoD rules
+  and violation evaluation, access review campaigns, role/permission
+  simulation (migration 132), RBAC document
+  upload/list/save-analysis/delete backed by migration 133
+  (`rbac_documents`), and row-level security for all four new tables
+  (migration 134).
+- New `rbac_analysis` AI feature added inline to the `llmService.js` monolith
+  (this repo has no split `services/ai/features/` layout), plus
+  `POST /api/v1/ai/rbac-analysis`.
+- Frontend `dashboard/access-governance` page with five tabs; the Import & AI
+  tab now lives in `components/accessGovernance/ImportAiTab.tsx`.
+- QA/tests: `qa-rbac-mega-test.js` section 22 (26 assertions), a Jest unit
+  suite for the service (13 tests), and schema tests.
+
+## Discrepancies found and fixed
+
+1. **No guide existed for this module.** `README.md` carried a single
+   Access Governance bullet, but there was no page explaining the permission
+   model, the campaign lifecycle, or the AI import flow — while the sibling
+   repo shipped a full guide. Added `docs/ACCESS_GOVERNANCE.md` (flat in
+   `docs/`, matching this repo's layout — it has no `docs/guides/`
+   subdirectory), adapted for this repo: migration numbers 132/133/134 rather
+   than 126/127/128, and the supported upload formats corrected to
+   PDF/DOCX/TXT/MD/CSV since this repo's
+   `VALID_DOCUMENT_EXTENSIONS` does not include `.doc`.
+2. **`GET /entitlements` became paginated in this pass** (`page`/`limit`,
+   default 50, max 100), with aggregates deliberately staying org-wide rather
+   than per-page. Documented in the new guide's Entitlements section.
+3. **The self-review restriction needed to be stated precisely.** Deciding a
+   review item requires `access_governance.manage`, which ships granted only
+   to `admin`; admins hold the `*` wildcard, which `requireSod` treats as an
+   authorized override. So the self-review block does not fire in a default
+   configuration — an admin can certify their own item, recorded with
+   `selfReviewOverride: true`. The guide says this plainly rather than
+   implying the control is absolute, because an assessor reading an AC-5
+   narrative would otherwise be misled.
+
+## Other fixes in this pass (no doc impact)
+
+- Upload failures now return 400 instead of 500. This repo was worse than the
+  sibling here: its multer `fileFilter` set no status at all, so *both*
+  unsupported file types and oversize files fell through to the generic 500
+  handler. Added a router-level `multer.MulterError` handler.
+- `createCampaign` batches its per-user inserts into a single `UNNEST`
+  statement instead of one round trip per user.
+- `simulateAccess` dedupes `roleIds` before validating them.
+- The Jest `decrypt` mock was an identity function, so it passed whether or
+  not the service decrypted `users.email`. Now returns distinguishable
+  plaintext, with three new `getEntitlementReport` tests pinning decryption
+  and pagination. Verified by reverting each fix and confirming failure.
+- `dashboard/access-governance/page.tsx` went from 1179 to 844 lines by
+  extracting the Import & AI tab into `components/accessGovernance/`. An
+  earlier note in this branch claimed no per-feature component subfolder
+  precedent existed in this repo; that was wrong — `components/rmf/` and
+  `components/ai/` both do.
+
+## Docs re-confirmed accurate, no change needed
+
+- `README.md` — the Access Governance bullet added earlier in this branch
+  still accurately describes the shipped feature set (entitlement reporting,
+  SoD rules, campaigns, simulator, AI-assisted RBAC import). No revision
+  needed; pagination is an API detail below README altitude.
+- `RELEASE_NOTES.md` — the two `## [Unreleased]` entries remain accurate.
+- `docs/FEDRAMP_DEPLOYMENT_GUIDE.md` — its "Network Topology" and database
+  sections are unaffected by this diff.
+
+## Out of scope (pre-existing, unrelated to this PR)
+
+- Migration 134's RLS policies — like migration 105's before them — are inert
+  when the app connects to Postgres as a superuser, since PostgreSQL always
+  exempts `rolsuper`/`rolbypassrls` roles and `FORCE ROW LEVEL SECURITY` only
+  removes the table owner's exemption. Now documented in the migration header.
+  Whether the deployment's `DATABASE_URL` uses a non-superuser role is an
+  infrastructure question outside this PR.
+- `docs/` has no index listing available guides, so the new
+  `ACCESS_GOVERNANCE.md` is discoverable only by directory listing. Consistent
+  with the other 14 docs there; not introduced here.
+
+## Conclusion
+
+The main documentation gap was the absence of any guide for this module in
+this repo, now closed with a ported and repo-adapted `ACCESS_GOVERNANCE.md`.
+`README.md` and `RELEASE_NOTES.md` were already accurate and needed no
+correction. The guide documents the pagination behavior added in this pass and
+states the real scope of the self-review control rather than overstating it.
